Files worked on sep5-20205

+++ version2/upload_stock.py
@@ -4,7 +4,6 @@ import config
 import pandas as pd
 
 import tempfile
-import asyncio
 import os
 import aiohttp
 import aiofiles
@@ -20,24 +19,31 @@ gemini = genai.Client()
 ZILLIZ_ENDPOINT = os.environ['ZILLIZ_ENDPOINT']
 ZILLIZ_TOKEN = os.environ['ZILLIZ_TOKEN']
 
-
+vs_id_var = "ITEM_CODE"
+excel_id_var = "ITEM CODE"
+excel_item_name_var = "ITEM NAME"
     
-async def create_records(records, item_codes):
+    
+async def create_records(embs, item_codes):
+    records = [{vs_id_var : item_codes[i], "embedding" : emb} for i, emb in enumerate(embs)]
+    records_ls = [records[i:i+100] for i in range(0, len(records), 100)]
+    
     url = ZILLIZ_ENDPOINT + "/v2/vectordb/entities/insert"
     headers = {
         "Authorization" : ZILLIZ_TOKEN,
         "Content-Type" : "application/json"
     }
-    payload = {
-        "collectionName": "treeyaa_stock_db_version2",
-        "data" : records
-    }
-    async with aiohttp.ClientSession() as session:
-        response = await session.post(url, headers = headers, json = payload)
-    if response.status == 200:
+    try:
+        async with aiohttp.ClientSession() as session:
+            for records_sub_ls in records_ls:
+                payload = {
+                    "collectionName": "treeyaa_vector_store",
+                    "data" : records_sub_ls
+                }
+                await session.post(url, headers = headers, json = payload)
         print(f"Created records {item_codes} in vector store successfully")
-    else:
-        print("Failed to create records in vector store. See the below error.\n\n", response)
+    except Exception as e:
+        print("Failed to create records in vector store. See the below error.\n\n", e)
     
 async def delete_records(item_codes):
     url = ZILLIZ_ENDPOINT + "/v2/vectordb/entities/delete"
@@ -46,8 +52,8 @@ async def delete_records(item_codes):
         "Content-Type" : "application/json"
     }
     payload = {
-        "collectionName": "treeyaa_stock_db_version2",
-        "filter": f"ITEM_CODE in {item_codes}"
+        "collectionName": "treeyaa_vector_store",
+        "filter": f"{vs_id_var} in {item_codes}"
         }
     async with aiohttp.ClientSession() as session:
         response = await session.post(url, headers = headers, json = payload)
@@ -60,21 +66,24 @@ async def delete_records(item_codes):
     
 async def generate_embeddings(tamil_names : list):
     contents = [types.Content(role = "user", parts = [types.Part.from_text(text = tamil_name)]) for tamil_name in tamil_names]
+    contents_ls = [contents[i:i+100] for i in range(0, len(contents), 100)]
+    embs = []
     try:
-        result = await gemini.aio.models.embed_content(
-                model = "gemini-embedding-001",
-                contents = contents)
-        emb = [emb.values for emb in result.embeddings]
+        for content_sub_ls in contents_ls:
+            result = await gemini.aio.models.embed_content(
+                    model = "gemini-embedding-001",
+                    contents = content_sub_ls)
+            emb = [emb.values for emb in result.embeddings]
+            embs += emb
         print("Successfully generated EMBEDDINGS")
-        return emb
+        return embs
     except Exception as e:
-        print("Failed to GENERATE EMBEDDINGS")
+        print("Failed to GENERATE EMBEDDINGS. Below is the error:\n\n", e)
     
 async def download_stock(link, file_path):
     try:
         async with aiohttp.ClientSession() as session:
             async with session.get(link) as response:
-                print(response.content_type)
                 data = await response.content.read()
                 async with aiofiles.open(file_path, mode = "wb") as f:
                     await f.write(data)
@@ -83,7 +92,6 @@ async def download_stock(link, file_path):
         print(f"Failed to download the stock excel file. Below is the error:\n\n{e}")
 
 async def generate_tamil_names(tanglish_names):
-    print(tanglish_names)
     contents = json.dumps({"items" : tanglish_names})
     try:
         response = await gemini.aio.models.generate_content(
@@ -92,7 +100,6 @@ async def generate_tamil_names(tanglish_names):
             config = config.ttt.config_2,
         )
         response = response.text
-        print(response)
         response = json.loads(response)
         print("Generated TAMIL names successfully")
         return response['items']
@@ -118,15 +125,15 @@ async def get_vector_store_ids():
         "Content-Type" : "application/json"
     }
     payload = {
-        "collectionName": "treeyaa_stock_db_version2",
-        "filter": "",
-        "output_fields" : ["ITEM_CODE"]
-        }
+        "collectionName" : "treeyaa_vector_store",
+        "outputFields" : ["ITEM_CODE"],
+        "limit" : 1000
+    }
     async with aiohttp.ClientSession() as session:
         response = await session.post(url, headers = headers, json = payload)
         res_obj = await response.json()
     res_obj = res_obj['data']
-    ids = [dic['ITEM_CODE'] for dic in res_obj]
+    ids = [dic[vs_id_var] for dic in res_obj]
     return ids
     
 async def upload_stock(link):
@@ -134,20 +141,35 @@ async def upload_stock(link):
         file_path = os.path.join(dir, "stock.xlsx")
         await download_stock(link, file_path)
         df = pd.read_excel(file_path, sheet_name = "OG", engine = "openpyxl")
-        vs_ids = await get_vector_store_ids()
-        to_update_item_codes = [id for id in df['ITEM CODE'] if id not in vs_ids]
-        to_delete_item_codes = [id for id in vs_ids if id not in df['ITEM CODE']]
-        print(to_delete_item_codes)
-        if to_update_item_codes:
-            tanglish_names = df.loc[df['ITEM CODE'].isin(to_update_item_codes), 'ITEM NAME'].tolist()
+        df[excel_id_var] = df[excel_id_var].astype(str)
+        
+        vs_item_codes = await get_vector_store_ids()
+        excel_item_codes = df[excel_id_var].to_list()
+        to_insert_item_codes = [item_code for item_code in excel_item_codes if item_code not in vs_item_codes]
+        to_delete_item_codes = [item_code for item_code in vs_item_codes if item_code not in excel_item_codes]
+        print("Excel item codes: ", excel_item_codes)
+        print("vs_item_codes: ", vs_item_codes)
+        print("to_insert_item_codes: ", to_insert_item_codes)
+        print("to_delete_item_codes: ", to_delete_item_codes)
+        
+        if to_insert_item_codes:
+            tanglish_names = df.loc[df[excel_id_var].isin(to_insert_item_codes), excel_item_name_var].tolist()
             tamil_names = await generate_tamil_names(tanglish_names)
             embs = await generate_embeddings(tamil_names)
-            records = [{"ITEM CODE" : to_update_item_codes[i], "embedding" : emb} for i, emb in enumerate(embs)]
-            await create_records(records, to_update_item_codes)
+            await create_records(embs, to_insert_item_codes)
+            
+            to_insert_items = df[df[excel_id_var].isin(to_insert_item_codes)].to_dict(orient = 'records')
+            insert_input = {"to_insert_items" : to_insert_items, "tamil_names" : tamil_names}
         else:
+            insert_input = None
             print("New ids not found so I didn't update the vector store")
-        '''if to_delete_item_codes:
+            
+        if to_delete_item_codes:
             await delete_records(to_delete_item_codes)
         else:
-            print("No record is deleted")'''
+            print("No record is deleted")
+            
+        update_input = df[~df[excel_id_var].isin(to_insert_item_codes)].to_dict(orient = 'records')
+        update_codes = df.loc[~df[excel_id_var].isin(to_insert_item_codes), excel_id_var].tolist()
+        await db_ops.upload_stock.run(insert_input, to_delete_item_codes, update_input, update_codes)
